rename sn_simulation.py sn_simclass.py
